requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.31.0
cloudscraper==1.2.66
colorama==0.4.6
cryptography==38.0.4
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1.3
django-allauth==0.51.0
django-celery-beat==2.4.0
django-timezone-field==5.0
exceptiongroup==1.0.4
gunicorn==20.1.0
h11==0.14.0
idna==3.4
kombu==5.2.4
lxml==4.9.2
memory-profiler==0.61.0
oauthlib==3.2.2
outcome==1.2.0
Pillow==9.3.0
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
PySocks==1.7.1
python-crontab==2.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7
redis==4.4.2
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
wsproto==1.2.0
cchardet==2.1.7
django-cors-headers==3.13.0